Revise the code based on Sourcery recommendations.
